requirements.txt
mysql-connector-python
ttkbootstrap